requirements.txt
numpy>=1.26
scipy>=1.11
pydantic>=2.5
python-dotenv>=1.0
fastapi>=0.109
uvicorn[standard]>=0.27
numpy-financial>=1.0
pytest>=7.4
hypothesis>=6.98
# numba>=0.59  # optional: faster IRR batch; pip may fail on very new Python—install if supported
